requirements.txt
Django==3.1.3
django-pandas==0.6.2
djangorestframework==3.12.2
dnspython==2.0.0
et-xmlfile==1.0.1
jdcal==1.4.1
mysqlclient==2.0.3
numpy==1.19.4
openpyxl==3.0.6
pandas==1.1.4
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2020.4
sqlparse==0.4.1